entity related packets and simplification of packet handling

+++ src/main/java/me/youhavetrouble/mobrrr/packet/Packet.java
@@ -2,6 +2,6 @@ package me.youhavetrouble.mobrrr.packet;
 
 public abstract class Packet {
 
-    public abstract int getId();
+    public abstract byte getId();
 
 }
